requirements.txt
Pillow==8.3.2
simplejson==3.17.5
pytest==6.2.3
